backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.20.0
pydantic>=2.11.0
pydantic-settings>=2.7.0
alembic>=1.14.0
python-dotenv>=1.0.1
anthropic>=0.40.0
twilio>=9.4.0
google-api-python-client>=2.155.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
apscheduler>=3.10.4
python-multipart>=0.0.20
httpx>=0.28.1
asyncpg>=0.30.0
